get ready base state controller and services

+++ src/main/java/uz/bookshop/repository/BooksRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import uz.bookshop.entity.Books;
 
 @Repository
-public interface BooksRepository extends JpaRepository<Books, Integer> {
+public interface BooksRepository extends JpaRepository<Books, Long> {
 }
