+++ src/main/java/uz/bookshop/repository/UserPurchasesRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import uz.bookshop.entity.UsersPurchases;
 
 @Repository
-public interface UserPurchasesRepository extends JpaRepository<UsersPurchases, Integer> {
+public interface UserPurchasesRepository extends JpaRepository<UsersPurchases, Long> {
 }
