+++ src/main/java/uz/bookshop/repository/UserRepository.java
@@ -5,7 +5,7 @@ import uz.bookshop.entity.Users;
 
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<Users, Integer> {
+public interface UserRepository extends JpaRepository<Users, Long> {
 
     Optional<Users> findByLogin(String login);
 }
